Added CacheAdapterBase and revised CacheItemFactory to simply be extensions.
